+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 | モジュール名 | Description | 最新のバージョン |
 | :-- | :-- | :-- |
-| KarteCore | イベントトラッキング機能を提供します。 | 2.20.0 |
+| KarteCore | イベントトラッキング機能を提供します。 | 2.21.0 |
 | KarteInAppMessaging | アプリ内メッセージ機能を提供します。 | 2.11.0 |
 | KarteRemoteNotification |  プッシュ通知の受信および効果測定機能を提供します。 | 2.7.0 |
 | KarteVariables | 設定値配信機能を提供します。 | 2.4.0 |
@@ -10,6 +10,15 @@
 | KarteCrashReporting  |  クラッシュイベントのトラッキング機能を提供します。 | 2.5.0 |
 | KarteUtilities |  KarteCore モジュール等が利用するUtility機能を提供します。通常直接参照する必要はありません。 | 3.7.0 |
 | KarteNotificationServiceExtension | リッチプッシュ通知機能を提供します。 | 1.0.0 |
+
+# Releases - 2022.07.29
+## Version 2.5.0
+
+### Core 2.21.0
+** 🎉 FEATURE**
+- WebView連携のための補助APIとして `UserSync.getUserSyncScript` を追加しました。
+  - 返されるスクリプトをWebViewで実行することで、`WKWebView`以外のWebViewに対してもユーザー連携が可能になります。
+  - これに伴い、クエリパラメータ連携API `UserSync.appendingQueryParameter` は非推奨になります。
 
 # Releases - 2022.06.28
 ## Version 2.4.0